Moves getCollectionByFullPath from graphql to schema-tools package.
