+++ packages/@tinacms/schema-tools/src/schema/TinaSchema.test.ts
@@ -1,6 +1,53 @@
 import { TinaSchema } from './TinaSchema';
 
 describe('TinaSchema', () => {
+  describe('getCollectionByFullPath', () => {
+    it('fetches correct collection with getCollectionByFullPath', async () => {
+      const schema = new TinaSchema({
+        version: { fullVersion: '', major: '', minor: '', patch: '' },
+        meta: { flags: [] },
+        collections: [
+          {
+            label: 'Test',
+            name: 'test',
+            path: 'content/test',
+            format: 'mdx',
+            fields: [
+              {
+                label: 'field',
+                type: 'string',
+                name: 'field',
+              },
+            ],
+          },
+          {
+            label: 'Test2',
+            name: 'test2',
+            path: 'content/test2',
+            format: 'mdx',
+            fields: [
+              {
+                label: 'field',
+                type: 'string',
+                name: 'field',
+              },
+            ],
+          },
+        ],
+      });
+
+      const collection1 = schema.getCollectionByFullPath(
+        'content/test/foobar.mdx'
+      );
+      expect(collection1?.name).toEqual('test');
+
+      const collection2 = schema.getCollectionByFullPath(
+        'content/test2/foobar.mdx'
+      );
+      expect(collection2?.name).toEqual('test2');
+    });
+  });
+
   describe('findReferencesFromCollection', () => {
     it('should find a single reference field', () => {
       const schema = new TinaSchema({
@@ -416,6 +463,7 @@ describe('TinaSchema', () => {
         path: '$.authorRefList[*].author',
       });
     });
+
     it('should walk nested fields within an object', () => {
       const schema = new TinaSchema({
         collections: [
@@ -571,6 +619,7 @@ describe('TinaSchema', () => {
         path: '$.metadata.priority',
       });
     });
+
     it('should walk a list of primitive fields', () => {
       const schema = new TinaSchema({
         collections: [
